requirements.txt
streamlit
python-dotenv
google-generativeai
faiss-cpu
pandas
numpy
PyPDF2
scikit-learn
